[SW-1476] Avoid duplication between mojo params and algo params
