+++ ml/src/main/scala/ai/h2o/sparkling/ml/params/AlgoParams.scala
@@ -14,57 +14,44 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-package org.apache.spark.ml.h2o.param
+package ai.h2o.sparkling.ml.params
 
-import java.util
-
+import hex.Model
+import hex.deeplearning.DeepLearningModel.DeepLearningParameters
+import hex.glm.GLMModel.GLMParameters
+import hex.tree.gbm.GBMModel.GBMParameters
+import hex.tree.xgboost.XGBoostModel.XGBoostParameters
+import org.apache.spark.ml.h2o.algos.H2OGridSearch
 import org.apache.spark.ml.param.{Param, Params}
 import org.json4s.JsonAST.{JArray, JInt}
 import org.json4s.jackson.JsonMethods.{compact, parse, render}
 import org.json4s.{JNull, JValue}
 import water.AutoBuffer
 
-import scala.collection.JavaConverters._
 
-class HyperParamsParam(parent: Params, name: String, doc: String, isValid: java.util.Map[String, Array[AnyRef]] => Boolean)
-  extends Param[java.util.Map[String, Array[AnyRef]]](parent, name, doc, isValid) {
+class AlgoParams(parent: Params, name: String, doc: String, isValid: Model.Parameters => Boolean)
+  extends Param[Model.Parameters](parent, name, doc, isValid) {
 
   def this(parent: Params, name: String, doc: String) =
     this(parent, name, doc, _ => true)
 
-  override def jsonEncode(value: java.util.Map[String, Array[AnyRef]]): String = {
-
+  override def jsonEncode(value: Model.Parameters): String = {
     val encoded: JValue = if (value == null) {
       JNull
     } else {
       val ab = new AutoBuffer()
-      ab.put1(value.size)
-      val it = value.entrySet().iterator()
-      while(it.hasNext){
-        val entry = it.next()
-        ab.putStr(entry.getKey)
-        //
-        //noinspection ComparingUnrelatedTypes
-        if(entry.getValue.isInstanceOf[util.ArrayList[Object]]){
-          val length = entry.getValue.asInstanceOf[util.ArrayList[_]].size()
-          val arrayList = entry.getValue.asInstanceOf[util.ArrayList[_]]
-          val arr = (0 until length).map(idx => arrayList.get(idx).asInstanceOf[AnyRef]).toArray
-          ab.putASer(arr)
-        } else {
-          ab.putASer(entry.getValue)
-        }
-      }
-
+      ab.putStr(value.algoName())
+      value.write(ab)
       val bytes = ab.buf()
       JArray(bytes.toSeq.map(JInt(_)).toList)
     }
     compact(render(encoded))
   }
 
-  override def jsonDecode(json: String): java.util.Map[String, Array[AnyRef]] = {
+  override def jsonDecode(json: String): Model.Parameters = {
     parse(json) match {
       case JNull =>
-        null
+        null.asInstanceOf[Model.Parameters]
       case JArray(values) =>
         val bytes = values.map {
           case JInt(x) =>
@@ -73,10 +60,20 @@ class HyperParamsParam(parent: Params, name: String, doc: String, isValid: java.
             throw new IllegalArgumentException(s"Cannot decode $json to Byte.")
         }.toArray
         val ab = new AutoBuffer(bytes)
-        val numParams = ab.get1()
-        (0 until numParams).map{ _ => (ab.getStr, ab.getASer[AnyRef](classOf[AnyRef]))}.toMap.asJava
+        val algoName = ab.getStr
+        val params = H2OGridSearch.SupportedAlgos.fromString(algoName).get  match {
+          case H2OGridSearch.SupportedAlgos.glm => new GLMParameters()
+          case H2OGridSearch.SupportedAlgos.gbm => new GBMParameters()
+          case H2OGridSearch.SupportedAlgos.deeplearning => new DeepLearningParameters()
+          case H2OGridSearch.SupportedAlgos.xgboost => new XGBoostParameters()
+
+          case _ => throw new RuntimeException("Not supported algorithm")
+
+        }
+        params.read(ab)
+        params.asInstanceOf[Model.Parameters]
       case _ =>
-        throw new IllegalArgumentException(s"Cannot decode $json to Map[String, Array[AnyRef]].")
+        throw new IllegalArgumentException(s"Cannot decode $json to a class for desired algo.")
     }
   }
 }
