+++ ml/src/main/scala/ai/h2o/sparkling/ml/params/H2OAlgoParamsHelper.scala
@@ -21,7 +21,6 @@ import java.util
 import com.google.common.base.CaseFormat
 import hex.Model.Parameters
 import org.apache.spark.h2o.utils.ReflectionUtils.api
-import org.apache.spark.ml.h2o.param.{NullableDoubleArrayParam, NullableStringArrayParam, NullableStringParam}
 import org.apache.spark.ml.param._
 
 import scala.reflect.ClassTag
