+++ ml/src/main/scala/org/apache/spark/ml/h2o/models/H2OMOJOModelBase.scala
@@ -17,13 +17,13 @@
 
 package org.apache.spark.ml.h2o.models
 
+import ai.h2o.sparkling.ml.params.H2OMOJOAlgoSharedParams
 import org.apache.spark.annotation.DeveloperApi
-import org.apache.spark.ml.h2o.param.H2OMOJOModelParams
 import org.apache.spark.sql.types.{StructField, StructType}
 import org.apache.spark.ml.{Model => SparkModel}
 
 abstract class H2OMOJOModelBase[T <: H2OMOJOModelBase[T]] extends SparkModel[T]
-  with H2OMOJOModelParams with HasMojoData with H2OMOJOWritable with H2OMOJOFlattenedInput {
+  with H2OMOJOAlgoSharedParams with HasMojoData with H2OMOJOWritable with H2OMOJOFlattenedInput {
 
   protected def getPredictionSchema(): Seq[StructField]
